requirements.txt
aksharamukha==2.2.2
fastapi==0.115.0
pydantic==2.9.2
pydantic_core==2.23.4
uvicorn==0.31.0
ai4bharat-transliteration-1.1.3 
alabaster-1.0.0 
antlr4-python3-runtime-4.8 
babel-2.16.0 
bitarray-3.0.0 
click-8.1.7 
colorama-0.4.6 
cython-3.0.11 
docutils-0.21.2 
fairseq-0.12.2 
flask-cors-5.0.0 
gevent-24.11.1 
hydra-core-1.0.7 
imagesize-1.4.1 
indic-nlp-library-0.92 
mock-5.1.0 
morfessor-2.0.6 
nvidia-cublas-cu12-12.4.5.8 
nvidia-cuda-cupti-cu12-12.4.127 
nvidia-cuda-nvrtc-cu12-12.4.127 
nvidia-cuda-runtime-cu12-12.4.127 
nvidia-cufft-cu12-11.2.1.3 
nvidia-curand-cu12-10.3.5.147 
nvidia-cusolver-cu12-11.6.1.9 
nvidia-cusparse-cu12-12.3.1.170 
nvidia-nccl-cu12-2.21.5 
nvidia-nvjitlink-cu12-12.4.127 
nvidia-nvtx-cu12-12.4.127 
omegaconf-2.0.6 
portalocker-3.0.0 
progressbar2-4.5.0 
pydload-1.0.9 
python-utils-3.9.1 
sacrebleu-2.4.3 
sacremoses-0.1.1 
snowballstemmer-2.2.0 
sphinx-8.1.3 
sphinx-argparse-0.5.2 
sphinx-rtd-theme-3.0.2 
sphinxcontrib-applehelp-2.0.0 
sphinxcontrib-devhelp-2.0.0 
sphinxcontrib-htmlhelp-2.1.0 
sphinxcontrib-jquery-4.1 
sphinxcontrib-jsmath-1.0.1 
sphinxcontrib-qthelp-2.0.0 
sphinxcontrib-serializinghtml-2.0.0 
sympy-1.13.1 
tabulate-0.9.0 
tensorboardX-2.6.2.2 
tomli-2.2.1 
torch-2.5.1 
torchaudio-2.5.1 
triton-3.1.0 
ujson-5.10.0 
urduhack-0.1.4 
zope.event-5.0 
zope.interface-7.2